Added PlayApp view state persistence
Issue from  Xyct:

This works perfectly with a showStartingProgress variable, but when isList is changed during starting, showStartingProgress is reset to false.

I then decide to rely on app.isStarting and update showStartingProgress from it, but something I don't understand happens: app.isStarting also resets to false. This not only lead to the not showing of the progress view when switching between List view and Grid view, but also makes the app can be launched multiple times during launching.